Update metadata and datapoint info for hledger
The tests for hledger were definitely either end-to-end or integration
tests as they called the hledger CLI interface to test one specific
thing. Even though the test itself was aimed at testing a specific
functionality or bug, the test was still implemented dependently on
everything else.

Additionally, I believe the dockerfiles could be changed to allow for
running with GHC 8.10.7 on all the hledger bugs